requirements.txt
asyncpraw
discord.py
python-dotenv
Flask
aiohttp
